Fix environment on Windows.
Installing shapely and leidenalg via conda resulted in conda complaining
that it could not resolve the environment; using pip instead solved this
problem. The current enviroment works on Windows with conda and the
libmamba solver, *if* basicpy is commented out and installed manually
(together with jax) afterwards, in the new conda environment.